Move MessageGenerator up to Channel level. So it can be reusable for other different channels
